+++ src/test/java/com/googongill/aditory/controller/UserControllerTest.java
@@ -6,8 +6,8 @@ import com.googongill.aditory.controller.dto.user.SignupRequest;
 import com.googongill.aditory.domain.User;
 import com.googongill.aditory.repository.UserRepository;
 import com.googongill.aditory.security.jwt.TokenProvider;
-import com.googongill.aditory.security.jwt.auth.PrincipalDetails;
-import com.googongill.aditory.security.jwt.auth.PrincipalDetailsService;
+import com.googongill.aditory.security.jwt.user.PrincipalDetails;
+import com.googongill.aditory.security.jwt.user.PrincipalDetailsService;
 import com.googongill.aditory.service.UserService;
 import com.googongill.aditory.service.dto.user.UserTokenResult;
 import com.googongill.aditory.service.dto.user.SignResult;
